# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Healthcare System API",
        "description": "API for Smart Healthcare System",
        "contact": {
            "email": "support@smart-healthcare.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "[AMAN] Get user profile",
                "description": "Get current authenticated user profile",
                "operationId": "getUserProfile",
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfileResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserProfile": {
                "title": "User Profile",
                "description": "User profile response",
                "properties": {
                    "id_pengguna": {
                        "type": "integer"
                    },
                    "nama_lengkap": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "pasien",
                            "dokter",
                            "admin"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "UserProfileResponse": {
                "title": "User Profile Response",
                "description": "Response untuk user profile",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/UserProfile"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}